将DataQuery重命名为SqlQuery，为增加MemQuery等做准备
