+++ src/main/java/cn/cerc/jdb/core/ITableOperator.java
@@ -1,11 +1,5 @@
 package cn.cerc.jdb.core;
 
-public interface ITableOperator {
-
-	public boolean insert(Record record);
-
-	public boolean update(Record record);
-
-	public boolean delete(Record record);
-
+@Deprecated
+public interface ITableOperator extends IDataOperator {
 }
